Put classes into their own files
